added some tests and cascade types for the rdb

+++ src/main/java/dev/silverscapes/smh/repository/InsuranceRepository.java
@@ -3,6 +3,6 @@ package dev.silverscapes.smh.repository;
 import dev.silverscapes.smh.model.Insurance;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface InsuranceRepository extends JpaRepository<Insurance,Integer> {
+public interface InsuranceRepository extends JpaRepository<Insurance,Long> {
 
 }
